Refactoring Points and Vectors into its own files and adding Docs
