+++ src/core3d/mod.rs
@@ -1,1 +1,3 @@
+pub mod point;
 pub mod tuples;
+pub mod vector;
