+++ src/core3d/tuples.rs
@@ -1,21 +1,55 @@
+/// This Tuple in 3D (x,y,z) space is a 4 unit set (x,y,z,w) to allow passing in either point or vector
 pub type Tuple = [f32; 4];
 
+/// # 3D Coordinates
+/// ## Left Handed Coordinates (default setup here)
+/// With the y axis pointing up, and the x axis pointing to the right, the z axis can be defined to point away from 
+/// you.
+/// ## Right Handed Coordinates
+/// With the y axis pointing up, and the x axis pointing to the right, the z axis can be defined to point toward you.
 pub trait Coord3D {
+    /// Gets the first dimension (right) from the coordinate system
     #[must_use]
     fn x(&self) -> f32;
 
+    /// Gets the second dimension (up) from the coordinate system
     #[must_use]
     fn y(&self) -> f32;
 
+    /// Gets the third dimension (away) from the coordinate system
     #[must_use]
     fn z(&self) -> f32;
 
+    /// Gets the fourth element `w` (usually 1.0 for absolute points, 0.0 for vectors) which helps with matrix 
+    /// multiplication
     #[must_use]
     fn w(&self) -> f32;
 
+    /// Checks if the tuple is a point.
+    ///
+    /// Returns `true` if is a point. `false` otherwise
+    ///
+    /// ### Example
+    /// ```
+    /// use crate::rusty_ray_tracer::core3d::tuples::Tuple;
+    /// use crate::rusty_ray_tracer::core3d::tuples::Coord3D;
+    /// let tuple: Tuple = [1.23, 4.56, 7.89, 1.0];
+    /// assert!(tuple.is_point() == true);
+    /// ```
     #[must_use]
     fn is_point(&self) -> bool;
 
+    /// Checks if the tuple is a vector.
+    ///
+    /// Returns `true` if is a vector. `false` otherwise
+    ///
+    /// ### Example
+    /// ```
+    /// use crate::rusty_ray_tracer::core3d::tuples::Tuple;
+    /// use crate::rusty_ray_tracer::core3d::tuples::Coord3D;
+    /// let tuple: Tuple = [1.23, 4.56, 7.89, 0.0];
+    /// assert!(tuple.is_vector() == true);
+    /// ```
     #[must_use]
     fn is_vector(&self) -> bool;
 }
@@ -89,76 +123,5 @@ mod tests {
         let tuple: Tuple = [1.23, 4.56, 7.89, 0.0];
         assert!(tuple.is_vector() == true);
         assert!(tuple.is_point() == false);
-    }
-}
-
-pub type PointF32 = Tuple;
-
-// pub trait Point {
-#[must_use]
-pub fn new_point(x: f32, y: f32, z: f32) -> PointF32 {
-    [x, y, z, 1.0]
-}
-// }
-// impl Point for PointF32 {}
-
-#[cfg(test)]
-mod tests_point {
-    use super::*;
-
-    #[test]
-    fn assign_array() {
-        let point: PointF32 = [4.0, 3.0, 2.0, 1.0];
-        assert_eq!(4.0, point[0]);
-        assert_eq!(3.0, point[1]);
-        assert_eq!(2.0, point[2]);
-        assert_eq!(1.0, point[3]);
-        assert!(point.is_point() == true);
-        assert!(point.is_vector() == false);
-    }
-
-    #[test]
-    fn create_new() {
-        let point = new_point(1.0, 2.0, 3.0);
-        assert_eq!(1.0, point.x());
-        assert_eq!(2.0, point.y());
-        assert_eq!(3.0, point.z());
-        assert_eq!(1.0, point.w());
-        assert!(point.is_point() == true);
-        assert!(point.is_vector() == false);
-    }
-}
-
-pub type VectorF32 = Tuple;
-
-#[must_use]
-pub fn new_vector(x: f32, y: f32, z: f32) -> VectorF32 {
-    [x, y, z, 0.0]
-}
-
-#[cfg(test)]
-mod tests_vector {
-    use super::*;
-
-    #[test]
-    fn assign_array() {
-        let vector: VectorF32 = [3.0, 2.0, 1.0, 0.0];
-        assert_eq!(3.0, vector[0]);
-        assert_eq!(2.0, vector[1]);
-        assert_eq!(1.0, vector[2]);
-        assert_eq!(0.0, vector[3]);
-        assert!(vector.is_vector() == true);
-        assert!(vector.is_point() == false);
-    }
-
-    #[test]
-    fn create_new() {
-        let vector = new_vector(1.0, 2.0, 3.0);
-        assert_eq!(1.0, vector.x());
-        assert_eq!(2.0, vector.y());
-        assert_eq!(3.0, vector.z());
-        assert_eq!(0.0, vector.w());
-        assert!(vector.is_vector() == true);
-        assert!(vector.is_point() == false);
     }
 }
